refactor: moved the types into a types folder + moved function that did not need to be exported into it's respective folders
