+++ package/src/use-form-assist.tsx
@@ -1,10 +1,85 @@
-import {useForm} from "./use-form"
-import {useValidation} from "./use-validation"
-import type {TValidation} from "./use-validation"
+import * as React from "react"
+import type * as TForm from "./types"
 
+function useValidation<TValues>(
+  _formValues: TValues,
+  validationOptions: TForm.TValidation<TValues>
+) {
+  const [validation] = React.useState<TForm.TValidation<TValues>>(
+    validationOptions
+  )
+
+  return validation
+}
+
+function useForm<TValues>(initialValues: TValues) {
+  function setInitialErrors() {
+    let errors: TForm.Errors<TValues> = {}
+    for (const key of Object.keys(initialValues)) {
+      errors = {...errors, [key]: ""}
+    }
+    return errors
+  }
+
+  function setInitialTouched({isTouched}: {isTouched: boolean}) {
+    let touched: TForm.Touched<TValues> = {}
+    for (const key of Object.keys(initialValues)) {
+      touched = {...touched, [key]: isTouched}
+    }
+    return touched
+  }
+
+  const [values, setValues] = React.useState<TValues>(initialValues)
+  const [errors, setErrors] = React.useState<TForm.Errors<TValues>>(() =>
+    setInitialErrors()
+  )
+  const [touched, setTouched] = React.useState<TForm.Touched<TValues>>(() =>
+    setInitialTouched({isTouched: false})
+  )
+
+  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
+    let value
+    if (e.target.type === "checkbox") {
+      value = e.target.checked
+    } else {
+      value = e.target.value
+    }
+    setValues({...values, [e.target.name]: value})
+  }
+
+  function handleBlur(e: React.FocusEvent<HTMLInputElement>) {
+    setTouched({...touched, [e.target.name]: true})
+  }
+
+  function resetForm() {
+    setValues(initialValues)
+    setErrors(setInitialErrors())
+    setTouched(setInitialTouched({isTouched: false}))
+  }
+
+  function touchAllFields() {
+    setTouched(setInitialTouched({isTouched: true}))
+  }
+
+  const formHelpers = {
+    values,
+    setValues,
+    errors,
+    setErrors,
+    touched,
+    handleChange,
+    handleBlur,
+    resetForm,
+    touchAllFields,
+  }
+
+  return formHelpers
+}
+
+// This hook will be used by the users, to call both the useForm and useValidaion hook togeher
 function useFormAssist<TValues>(
   initialValues: TValues,
-  validationOptions: TValidation<TValues> = {}
+  validationOptions: TForm.TValidation<TValues> = {}
 ) {
   const formHelpers = useForm(initialValues)
   const validation = useValidation(initialValues, validationOptions)
