--- package/src/use-validation.tsx
@@ -1,22 +1,0 @@
-import * as React from "react"
-
-type TValidation<TValues> = {
-  [key in keyof TValues]?: {
-    required?: boolean
-    maxCharacters?: number
-    minCharacters?: number
-    email?: boolean
-  }
-}
-
-function useValidation<TValues>(
-  _formValues: TValues,
-  validationOptions: TValidation<TValues>
-) {
-  const [validation] = React.useState<TValidation<TValues>>(validationOptions)
-
-  return validation
-}
-
-export type {TValidation}
-export {useValidation}
